+++ frontend/src/pages/Dashboard.tsx
@@ -1,688 +1,33 @@
-import { useCallback, useEffect, useMemo, useRef, useState } from "react";
-import { useNavigate } from "react-router-dom";
 import DisasterResponseAssistant from "@/components/features/DisasterResponseAssistant.tsx";
 import AppSidebar from "@/components/layout/AppSidebar";
 import Container from "@/components/layout/Container";
 import Footer from "@/components/layout/Footer.tsx";
 import Header from "@/components/layout/Header";
-import { Button } from "@/components/ui/Button";
-import Item from "@/components/ui/Item";
-import Pagination from "@/components/ui/Pagination";
 import { SpinnerEmpty } from "@/components/ui/SpinnerEmpty";
-import {
-	Select,
-	SelectContent,
-	SelectGroup,
-	SelectItem,
-	SelectLabel,
-	SelectTrigger,
-	SelectValue,
-} from "@/components/ui/select";
-import { Switch } from "@/components/ui/switch";
-import {
-	Tooltip,
-	TooltipContent,
-	TooltipProvider,
-	TooltipTrigger,
-} from "@/components/ui/tooltip";
-import { DAMAGE_COLOR_HEX } from "@/constants/app";
-import {
-	ApiError,
-	type BuildingFeature,
-	type BuildingStatsResponse,
-	fetchBuildingStatsForDisaster,
-	fetchBuildingsForDisaster,
-	fetchDisasters,
-	fetchImagePairs,
-	type ImagePairFeature,
-	resolveImageUrl,
-} from "@/utils/hazardlyApi";
-
-type DamageLevel = "no-damage" | "minor-damage" | "major-damage" | "destroyed";
-
-type NormalizedDamage = DamageLevel | "un-classified";
-
-type BuildingListItem = {
-	id: number;
-	uid: string;
-	address?: string | null;
-	image_pair_id: number;
-	xbd_id: number;
-	disaster_name: string;
-	actual_damage: NormalizedDamage;
-	predicted_damage?: NormalizedDamage | null;
-	is_correct?: boolean | null;
-	created_at?: string | null;
-	pre_image_path?: string | null;
-	post_image_path?: string | null;
-};
-
-type ImagePairRow = {
-	id: number;
-	xbd_id: number;
-	pre_image_path?: string | null;
-	post_image_path?: string | null;
-	totalBuildings: number;
-	correctCount: number;
-	incorrectCount: number;
-	comparedCount: number;
-	accuracyPct: string | null;
-};
-
-type ActiveSection = "overview" | "buildings" | "image-pairs";
-
-const CONFUSION_LABELS: DamageLevel[] = [
-	"no-damage",
-	"minor-damage",
-	"major-damage",
-	"destroyed",
-];
-
-const PAGE_SIZE = 25;
-const DAMAGE_FILTERS: Array<{ key: string; label: string }> = [
-	{ key: "all", label: "All" },
-	{ key: "no-damage", label: "No Damage" },
-	{ key: "minor-damage", label: "Minor Damage" },
-	{ key: "major-damage", label: "Major Damage" },
-	{ key: "destroyed", label: "Destroyed" },
-];
-const DAMAGE_SYNONYMS: Record<string, NormalizedDamage> = {
-	"no-damage": "no-damage",
-	"no-damages": "no-damage",
-	"minor-damage": "minor-damage",
-	"minor-damages": "minor-damage",
-	"major-damage": "major-damage",
-	"major-damages": "major-damage",
-	destroyed: "destroyed",
-	destroy: "destroyed",
-	"un-classified": "un-classified",
-	unclassified: "un-classified",
-	unknown: "un-classified",
-	uncertain: "un-classified",
-};
-
-const prettyLabel = (value: string): string =>
-	value
-		.split("-")
-		.map((part) => part.charAt(0).toUpperCase() + part.slice(1))
-		.join(" ");
-
-const normalizeDamage = (raw?: string | null): NormalizedDamage => {
-	if (!raw) return "un-classified";
-
-	const normalized = raw
-		.toLowerCase()
-		.trim()
-		.replace(/[_\s]+/g, "-")
-		.replace(/-+/g, "-");
-
-	const mappedDamage = DAMAGE_SYNONYMS[normalized];
-	if (mappedDamage) return mappedDamage;
-
-	if (DAMAGE_FILTERS.some((item) => item.key === normalized)) {
-		return normalized as NormalizedDamage;
-	}
-
-	return "un-classified";
-};
-
-const toPct = (value: number, total: number): string => {
-	if (total <= 0) return "0.0";
-	return ((value / total) * 100).toFixed(1);
-};
-
-const createEmptyConfusionMatrix = (): Record<
-	DamageLevel,
-	Record<DamageLevel, number>
-> => ({
-	"no-damage": {
-		"no-damage": 0,
-		"minor-damage": 0,
-		"major-damage": 0,
-		destroyed: 0,
-	},
-	"minor-damage": {
-		"no-damage": 0,
-		"minor-damage": 0,
-		"major-damage": 0,
-		destroyed: 0,
-	},
-	"major-damage": {
-		"no-damage": 0,
-		"minor-damage": 0,
-		"major-damage": 0,
-		destroyed: 0,
-	},
-	destroyed: {
-		"no-damage": 0,
-		"minor-damage": 0,
-		"major-damage": 0,
-		destroyed: 0,
-	},
-});
-
-const buildStatsFromFeatures = (
-	features: BuildingFeature[],
-): BuildingStatsResponse => {
-	const byDamage: Record<string, number> = {
-		"no-damage": 0,
-		"minor-damage": 0,
-		"major-damage": 0,
-		destroyed: 0,
-	};
-
-	for (const feature of features) {
-		const key = normalizeDamage(feature.properties.actual_damage);
-		if (key === "un-classified") continue;
-		byDamage[key] = (byDamage[key] ?? 0) + 1;
-	}
-
-	const total = features.length;
-
-	return {
-		total,
-		no_damage: byDamage["no-damage"] ?? 0,
-		by_damage: byDamage,
-	};
-};
-
-const resolveStorageUrl = (path?: string | null): string | null =>
-	path ? resolveImageUrl(path) : null;
+import DashboardBuildingsSection from "./dashboard/DashboardBuildingsSection";
+import DashboardImagePairsSection from "./dashboard/DashboardImagePairsSection";
+import DashboardOverviewSection from "./dashboard/DashboardOverviewSection";
+import { useDashboardData } from "./dashboard/useDashboardData";
 
 export default function Dashboard() {
-	const navigate = useNavigate();
-	const [loadingStats, setLoadingStats] = useState(true);
-	const [loadingBuildings, setLoadingBuildings] = useState(false);
-	const [loadingImagePairs, setLoadingImagePairs] = useState(false);
-	const [error, setError] = useState<string | null>(null);
-	const [activeSection, setActiveSection] = useState<ActiveSection>("overview");
-	const [activeDamageFilter, setActiveDamageFilter] = useState<string>("all");
-	const [showCorrectOnly, setShowCorrectOnly] = useState(false);
-	const [selectedDisasterId, setSelectedDisasterId] = useState<number | null>(
-		null,
-	);
-	const [selectedDisasterName, setSelectedDisasterName] = useState<
-		string | null
-	>(null);
-	const [allBuildingsCache, setAllBuildingsCache] = useState<
-		BuildingFeature[] | null
-	>(null);
-	const [imagePairMap, setImagePairMap] = useState<
-		Map<
-			number,
-			{
-				xbd_id: number;
-				pre_image_path?: string | null;
-				post_image_path?: string | null;
-			}
-		>
-	>(new Map());
-	const [stats, setStats] = useState<BuildingStatsResponse>({
-		total: 0,
-		no_damage: 0,
-		by_damage: {},
-	});
-	const [rows, setRows] = useState<BuildingListItem[]>([]);
-	const [page, setPage] = useState(1);
-	const [totalPages, setTotalPages] = useState(1);
-	const [totalItems, setTotalItems] = useState(0);
-	const [imagePairsPage, setImagePairsPage] = useState(1);
-	const imagePairMapRef = useRef(imagePairMap);
-	const allBuildingsCacheRef = useRef(allBuildingsCache);
-	const hasLoadedInitialStatsRef = useRef(false);
-
-	imagePairMapRef.current = imagePairMap;
-	allBuildingsCacheRef.current = allBuildingsCache;
-
-	const ensureImagePairs = useCallback(async (disasterId: number) => {
-		if (imagePairMapRef.current.size > 0) return imagePairMapRef.current;
-		const data = await fetchImagePairs(disasterId);
-		const next = new Map<
-			number,
-			{
-				xbd_id: number;
-				pre_image_path?: string | null;
-				post_image_path?: string | null;
-			}
-		>();
-		for (const feature of data.features ?? []) {
-			const props = feature.properties as ImagePairFeature["properties"] & {
-				id?: number;
-			};
-			if (typeof props.id !== "number") continue;
-			next.set(props.id, {
-				xbd_id: feature.properties.xbd_id,
-				pre_image_path: feature.properties.pre_image_path ?? null,
-				post_image_path: feature.properties.post_image_path ?? null,
-			});
-		}
-		imagePairMapRef.current = next;
-		setImagePairMap(next);
-		return next;
-	}, []);
-
-	const ensureAllBuildings = useCallback(async (disasterId: number) => {
-		if (allBuildingsCacheRef.current) return allBuildingsCacheRef.current;
-		const data = await fetchBuildingsForDisaster(disasterId);
-		const features = data.features ?? [];
-		allBuildingsCacheRef.current = features;
-		setAllBuildingsCache(features);
-		return features;
-	}, []);
-
-	useEffect(() => {
-		if (hasLoadedInitialStatsRef.current) return;
-		hasLoadedInitialStatsRef.current = true;
-
-		const loadStats = async () => {
-			try {
-				setLoadingStats(true);
-				setError(null);
-
-				const disasters = await fetchDisasters();
-				if (disasters.length === 0) {
-					setError("No disasters available.");
-					setLoadingStats(false);
-					return;
-				}
-
-				const currentDisaster = disasters[0];
-				setSelectedDisasterId(currentDisaster.id);
-				setSelectedDisasterName(currentDisaster.name);
-
-				// Wait for BOTH the stats and the full features before turning off the loader.
-				const [allBuildingsResult, statsResult] = await Promise.allSettled([
-					ensureAllBuildings(currentDisaster.id),
-					(async () => {
-						try {
-							const stats = await fetchBuildingStatsForDisaster(
-								currentDisaster.id,
-							);
-							setStats(stats);
-						} catch (err) {
-							if (err instanceof ApiError && err.status === 404) {
-								const features = await ensureAllBuildings(currentDisaster.id);
-								setStats(buildStatsFromFeatures(features));
-								return;
-							}
-							throw err;
-						}
-					})(),
-				]);
-
-				if (allBuildingsResult.status === "rejected") {
-					setError(
-						allBuildingsResult.reason instanceof Error
-							? allBuildingsResult.reason.message
-							: "Unexpected dashboard error.",
-					);
-					return;
-				}
-				if (statsResult.status === "rejected") {
-					setError(
-						statsResult.reason instanceof Error
-							? statsResult.reason.message
-							: "Unexpected dashboard error.",
-					);
-					return;
-				}
-			} catch (err) {
-				setError(
-					err instanceof Error ? err.message : "Unexpected dashboard error.",
-				);
-			} finally {
-				setLoadingStats(false);
-			}
-		};
-
-		void loadStats();
-	}, [ensureAllBuildings]);
-
-	const predictionMetrics = useMemo(() => {
-		const confusionMatrix = createEmptyConfusionMatrix();
-
-		if (!allBuildingsCache) {
-			return {
-				correctCount: 0,
-				comparedCount: 0,
-				accuracyPct: "0.0",
-				confusionMatrix,
-				available: false,
-				matrixTotal: 0,
-				matrixMax: 0,
-			};
-		}
-
-		let correctCount = 0;
-		let comparedCount = 0;
-
-		for (const feature of allBuildingsCache) {
-			const actual = normalizeDamage(feature.properties.actual_damage);
-			const rawPredicted = feature.properties.predicted_damage;
-			if (rawPredicted == null) continue;
-
-			const predicted = normalizeDamage(rawPredicted);
-			if (actual === "un-classified" || predicted === "un-classified") continue;
-			comparedCount += 1;
-			if (actual === predicted) correctCount += 1;
-
-			confusionMatrix[actual][predicted] += 1;
-		}
-
-		let matrixMax = 0;
-		for (const actualLabel of CONFUSION_LABELS) {
-			for (const predictedLabel of CONFUSION_LABELS) {
-				const value = confusionMatrix[actualLabel][predictedLabel];
-				if (value > matrixMax) matrixMax = value;
-			}
-		}
-
-		const accuracyPct =
-			comparedCount > 0
-				? ((correctCount / comparedCount) * 100).toFixed(1)
-				: "0.0";
-
-		return {
-			correctCount,
-			comparedCount,
-			accuracyPct,
-			confusionMatrix,
-			available: comparedCount > 0,
-			matrixTotal: comparedCount,
-			matrixMax,
-		};
-	}, [allBuildingsCache]);
-
-	useEffect(() => {
-		if (activeSection !== "buildings") return;
-		if (selectedDisasterId == null) return;
-
-		const loadBuildingsPage = async () => {
-			try {
-				setLoadingBuildings(true);
-				setError(null);
-
-				const pairs = await ensureImagePairs(selectedDisasterId);
-
-				const features = await ensureAllBuildings(selectedDisasterId);
-
-				const damageFiltered =
-					activeDamageFilter === "all"
-						? features
-						: features.filter(
-								(feature) =>
-									normalizeDamage(feature.properties.actual_damage) ===
-									activeDamageFilter,
-							);
-
-				const filtered = showCorrectOnly
-					? damageFiltered.filter((feature) => {
-							const actual = normalizeDamage(feature.properties.actual_damage);
-							const rawPredicted = feature.properties.predicted_damage;
-							if (rawPredicted == null) return false;
-
-							const predicted = normalizeDamage(rawPredicted);
-							if (actual === "un-classified" || predicted === "un-classified") {
-								return false;
-							}
-
-							return actual === predicted;
-						})
-					: damageFiltered;
-
-				const nextTotalItems = filtered.length;
-				const nextTotalPages = Math.max(
-					1,
-					Math.ceil(nextTotalItems / PAGE_SIZE),
-				);
-				const clampedPage = Math.min(page, nextTotalPages);
-
-				// Keep page in range when filters reduce total result size.
-				if (clampedPage !== page) {
-					setPage(clampedPage);
-					return;
-				}
-
-				const start = (clampedPage - 1) * PAGE_SIZE;
-				const pagedFeatures = filtered.slice(start, start + PAGE_SIZE);
-
-				const mapped: BuildingListItem[] = pagedFeatures.map((feature) => {
-					const prop = feature.properties;
-					const pair = pairs.get(prop.image_pair_id);
-
-					const actual = normalizeDamage(prop.actual_damage);
-					const predicted =
-						prop.predicted_damage == null
-							? null
-							: normalizeDamage(prop.predicted_damage);
-
-					return {
-						id: prop.id,
-						uid: prop.uid,
-						address: typeof prop.address === "string" ? prop.address : null,
-						disaster_name: selectedDisasterName ?? "Unknown",
-						image_pair_id: prop.image_pair_id,
-						xbd_id: pair?.xbd_id ?? -1,
-						actual_damage: actual,
-						predicted_damage: predicted,
-						is_correct: predicted === null ? null : actual === predicted,
-						created_at: prop.created_at ?? null,
-						pre_image_path: pair?.pre_image_path ?? null,
-						post_image_path: pair?.post_image_path ?? null,
-					};
-				});
-
-				setRows(mapped);
-				setTotalItems(nextTotalItems);
-				setTotalPages(nextTotalPages);
-			} catch (err) {
-				setError(
-					err instanceof Error
-						? err.message
-						: "Unexpected dashboard error while loading buildings.",
-				);
-				setRows([]);
-			} finally {
-				setLoadingBuildings(false);
-			}
-		};
-
-		void loadBuildingsPage();
-	}, [
-		activeSection,
-		selectedDisasterId,
-		page,
-		activeDamageFilter,
-		showCorrectOnly,
-		ensureAllBuildings,
-		ensureImagePairs,
-		selectedDisasterName,
-	]);
-
-	useEffect(() => {
-		if (activeSection !== "image-pairs") return;
-		if (selectedDisasterId == null) return;
-
-		const loadImagePairs = async () => {
-			try {
-				setLoadingImagePairs(true);
-				setError(null);
-				await Promise.all([
-					ensureImagePairs(selectedDisasterId),
-					ensureAllBuildings(selectedDisasterId),
-				]);
-			} catch (err) {
-				setError(
-					err instanceof Error
-						? err.message
-						: "Unexpected dashboard error while loading image pairs.",
-				);
-			} finally {
-				setLoadingImagePairs(false);
-			}
-		};
-
-		void loadImagePairs();
-	}, [activeSection, selectedDisasterId, ensureImagePairs, ensureAllBuildings]);
-
-	const overviewDamageRows = useMemo(() => {
-		const predictedByDamage: Record<string, number> = {
-			"no-damage": 0,
-			"minor-damage": 0,
-			"major-damage": 0,
-			destroyed: 0,
-		};
-
-		let predictedTotal = 0;
-		for (const feature of allBuildingsCache ?? []) {
-			const rawPredicted = feature.properties.predicted_damage;
-			if (rawPredicted == null) continue;
-			const key = normalizeDamage(rawPredicted);
-			if (key === "un-classified") continue;
-			predictedByDamage[key] = (predictedByDamage[key] ?? 0) + 1;
-			predictedTotal += 1;
-		}
-
-		return DAMAGE_FILTERS.filter((item) => item.key !== "all").map((item) => {
-			const actualCount = stats.by_damage[item.key] ?? 0;
-			const predictedCount = predictedByDamage[item.key] ?? 0;
-			const classKey = item.key as DamageLevel;
-			const actualTotalWithPrediction = CONFUSION_LABELS.reduce(
-				(sum, predictedLabel) =>
-					sum +
-					(predictionMetrics.confusionMatrix[classKey][predictedLabel] ?? 0),
-				0,
-			);
-			const diagonal =
-				predictionMetrics.confusionMatrix[classKey][classKey] ?? 0;
-			const classAccuracy =
-				actualTotalWithPrediction > 0
-					? `${toPct(diagonal, actualTotalWithPrediction)}%`
-					: "0.0%";
-			return {
-				key: item.key,
-				label: item.label,
-				actualCount,
-				actualPercentage: toPct(actualCount, stats.total),
-				predictedCount,
-				predictedPercentage: toPct(predictedCount, predictedTotal),
-				classAccuracy,
-				color: DAMAGE_COLOR_HEX[item.key] ?? "#9ca3af",
-			};
-		});
-	}, [stats, allBuildingsCache, predictionMetrics]);
-
-	const imagePairRows = useMemo(() => {
-		if (!allBuildingsCache || imagePairMap.size === 0)
-			return [] as ImagePairRow[];
-
-		const counts = new Map<
-			number,
-			{
-				total: number;
-				correct: number;
-				incorrect: number;
-				compared: number;
-			}
-		>();
-
-		for (const feature of allBuildingsCache) {
-			const pairId = feature.properties.image_pair_id;
-			const current = counts.get(pairId) ?? {
-				total: 0,
-				correct: 0,
-				incorrect: 0,
-				compared: 0,
-			};
-			current.total += 1;
-
-			const actual = normalizeDamage(feature.properties.actual_damage);
-			const rawPredicted = feature.properties.predicted_damage;
-			if (rawPredicted != null) {
-				const predicted = normalizeDamage(rawPredicted);
-				if (actual !== "un-classified" && predicted !== "un-classified") {
-					current.compared += 1;
-					if (actual === predicted) {
-						current.correct += 1;
-					} else {
-						current.incorrect += 1;
-					}
-				}
-			}
-
-			counts.set(pairId, current);
-		}
-
-		const rows = Array.from(imagePairMap.entries()).map(([id, pair]) => {
-			const metrics = counts.get(id) ?? {
-				total: 0,
-				correct: 0,
-				incorrect: 0,
-				compared: 0,
-			};
-			const accuracyPct =
-				metrics.compared > 0
-					? ((metrics.correct / metrics.compared) * 100).toFixed(1)
-					: null;
-			return {
-				id,
-				xbd_id: pair.xbd_id,
-				pre_image_path: pair.pre_image_path ?? null,
-				post_image_path: pair.post_image_path ?? null,
-				totalBuildings: metrics.total,
-				correctCount: metrics.correct,
-				incorrectCount: metrics.incorrect,
-				comparedCount: metrics.compared,
-				accuracyPct,
-			};
-		});
-
-		rows.sort((a, b) => a.xbd_id - b.xbd_id);
-		return rows;
-	}, [allBuildingsCache, imagePairMap]);
-
-	const imagePairTotalItems = imagePairRows.length;
-	const imagePairTotalPages = Math.max(
-		1,
-		Math.ceil(imagePairTotalItems / PAGE_SIZE),
-	);
-	const imagePairPageClamped = Math.min(imagePairsPage, imagePairTotalPages);
-	const imagePairPageRows = imagePairRows.slice(
-		(imagePairPageClamped - 1) * PAGE_SIZE,
-		imagePairPageClamped * PAGE_SIZE,
-	);
-
-	useEffect(() => {
-		if (imagePairsPage !== imagePairPageClamped) {
-			setImagePairsPage(imagePairPageClamped);
-		}
-	}, [imagePairsPage, imagePairPageClamped]);
-
-	const setDamageFilter = (filter: string) => {
-		setActiveDamageFilter(filter);
-		setPage(1);
-		setActiveSection("buildings");
-	};
+	const dashboard = useDashboardData();
 
 	return (
 		<div className="min-h-screen bg-background text-foreground">
 			<Header />
 			<Container className="w-full py-8">
-				<div className="flex flex-row gap-6 items-start">
+				<div className="flex flex-row items-start gap-6">
 					<AppSidebar
-						activeSection={activeSection}
-						onOverview={() => setActiveSection("overview")}
-						onBuildings={() => setActiveSection("buildings")}
-						onImagePairs={() => {
-							setActiveSection("image-pairs");
-							setImagePairsPage(1);
-						}}
-						disabled={loadingStats}
+						activeSection={dashboard.activeSection}
+						onOverview={dashboard.setOverviewSection}
+						onBuildings={dashboard.setBuildingsSection}
+						onImagePairs={dashboard.setImagePairsSection}
+						disabled={dashboard.loadingStats}
 					/>
 
-					<div className="space-y-6 min-w-0 flex-1 relative">
-						{loadingStats && (
-							<div className="absolute inset-0 z-50 flex items-center justify-center bg-background/60 backdrop-blur-sm rounded-xl">
+					<div className="relative min-w-0 flex-1 space-y-6">
+						{dashboard.loadingStats && (
+							<div className="absolute inset-0 z-50 flex items-center justify-center rounded-xl bg-background/60 backdrop-blur-sm">
 								<SpinnerEmpty
 									title="Loading Dashboard"
 									description="Syncing disaster data..."
@@ -691,522 +36,50 @@ export default function Dashboard() {
 							</div>
 						)}
 
-						{error && (
+						{dashboard.error && (
 							<div className="rounded-xl border border-destructive/40 bg-card p-4 text-sm text-destructive">
-								{error}
+								{dashboard.error}
 							</div>
 						)}
 
-						{activeSection === "overview" && (
-							<div
-								className={`${
-									loadingStats
-										? "opacity-20 pointer-events-none"
-										: "opacity-100 transition-opacity duration-300"
-								} flex flex-col gap-6`}
-							>
-								<div className="grid grid-cols-1 sm:grid-cols-2 lg:grid-cols-2 gap-4">
-									<div className="rounded-xl border border-border bg-card text-card-foreground p-4 flex min-h-[180px] flex-col">
-										<p className="text-sm text-muted-foreground">
-											Total Buildings
-										</p>
-										<p className="mt-2 text-4xl font-bold leading-none">
-											{stats.total}
-										</p>
-										<Button
-											variant="outline"
-											size="sm"
-											className="mt-auto"
-											onClick={() => setDamageFilter("all")}
-										>
-											View
-										</Button>
-									</div>
-
-									<div className="rounded-xl border border-border bg-card text-card-foreground p-4 flex min-h-[180px] flex-col">
-										<p className="text-sm text-muted-foreground">
-											Overall Accuracy
-										</p>
-										<p className="mt-2 text-4xl font-bold leading-none">
-											{predictionMetrics.available
-												? `${predictionMetrics.accuracyPct}%`
-												: "N/A"}
-										</p>
-										<div className="mt-auto space-y-1">
-											<p className="text-xs text-muted-foreground">
-												{predictionMetrics.correctCount} exact matches from{" "}
-												{predictionMetrics.comparedCount} predictions
-											</p>
-											<p className="text-[11px] text-muted-foreground">
-												Exact class match between ground truth and VLM output.
-											</p>
-										</div>
-									</div>
-								</div>
-
-								<div className="rounded-xl border border-border bg-card text-card-foreground p-5">
-									<div className="mb-4 flex items-center justify-between gap-3">
-										<h3 className="text-lg font-semibold">
-											Damage Distribution
-										</h3>
-										<span className="inline-flex items-center gap-1 text-[11px] text-muted-foreground">
-											<span
-												className="h-2 w-2 rounded-full"
-												style={{ backgroundColor: "#2563eb", opacity: 0.85 }}
-											/>
-											Predicted
-										</span>
-									</div>
-									<div className="space-y-4">
-										<TooltipProvider>
-											{overviewDamageRows.map((row) => (
-												<div key={row.key} className="space-y-1.5">
-													<div className="flex items-center justify-between text-sm">
-														<span className="font-medium">{row.label}</span>
-														<span className="text-xs text-muted-foreground">
-															{row.classAccuracy}
-														</span>
-													</div>
-
-													<Tooltip>
-														<TooltipTrigger asChild>
-															<div className="relative h-2.5 w-full cursor-help overflow-hidden rounded-full bg-muted">
-																<div
-																	className="absolute left-0 top-0 h-2.5 rounded-full"
-																	style={{
-																		width: `${row.actualPercentage}%`,
-																		backgroundColor: row.color,
-																		opacity: 0.75,
-																	}}
-																/>
-															</div>
-														</TooltipTrigger>
-														<TooltipContent side="top">
-															<p>
-																Actual: {row.actualCount} (
-																{row.actualPercentage}%)
-															</p>
-														</TooltipContent>
-													</Tooltip>
-
-													<Tooltip>
-														<TooltipTrigger asChild>
-															<div className="relative h-2.5 w-full cursor-help overflow-hidden rounded-full bg-muted">
-																<div
-																	className="absolute left-0 top-0 h-2.5 rounded-full"
-																	style={{
-																		width: `${row.predictedPercentage}%`,
-																		backgroundColor: "#2563eb",
-																		opacity: 0.85,
-																	}}
-																/>
-															</div>
-														</TooltipTrigger>
-														<TooltipContent side="top">
-															<p>
-																Predicted: {row.predictedCount} (
-																{row.predictedPercentage}%)
-															</p>
-														</TooltipContent>
-													</Tooltip>
-												</div>
-											))}
-										</TooltipProvider>
-									</div>
-								</div>
-
-								<div className="rounded-xl border border-border bg-card text-card-foreground p-5">
-									<div className="mb-4 flex items-end justify-between gap-3">
-										<div>
-											<h3 className="text-lg font-semibold">
-												Confusion Matrix Heatmap
-											</h3>
-											<p className="text-xs text-muted-foreground">
-												Rows = actual labels, columns = VLM predictions
-											</p>
-										</div>
-										<p className="text-xs text-muted-foreground">
-											{predictionMetrics.matrixTotal} predictions compared
-										</p>
-									</div>
-
-									{predictionMetrics.available ? (
-										<div
-											className={`overflow-x-auto ${loadingBuildings ? "opacity-50" : "opacity-100"}`}
-										>
-											<table className="w-full min-w-[760px] text-sm">
-												<thead>
-													<tr className="border-b border-border/60 text-left text-muted-foreground bg-muted/30">
-														<th className="px-2 py-2 font-medium">
-															Actual \ Predicted
-														</th>
-														{CONFUSION_LABELS.map((predictedLabel) => (
-															<th
-																key={`pred-${predictedLabel}`}
-																className="px-2 py-2 font-medium text-center"
-															>
-																{prettyLabel(predictedLabel)}
-															</th>
-														))}
-														<th className="px-2 py-2 font-semibold text-center text-foreground">
-															Total
-														</th>
-													</tr>
-												</thead>
-												<tbody>
-													{CONFUSION_LABELS.map((actualLabel) => {
-														const rowTotal = CONFUSION_LABELS.reduce(
-															(sum, predictedLabel) =>
-																sum +
-																(predictionMetrics.confusionMatrix[actualLabel][
-																	predictedLabel
-																] ?? 0),
-															0,
-														);
-
-														return (
-															<tr
-																key={`row-${actualLabel}`}
-																className="border-b border-border/70"
-															>
-																<td className="px-2 py-2 font-medium">
-																	{prettyLabel(actualLabel)}
-																</td>
-																{CONFUSION_LABELS.map((predictedLabel) => {
-																	const value =
-																		predictionMetrics.confusionMatrix[
-																			actualLabel
-																		][predictedLabel] ?? 0;
-																	const isDiagonal =
-																		actualLabel === predictedLabel;
-																	const intensity =
-																		predictionMetrics.matrixMax > 0
-																			? value / predictionMetrics.matrixMax
-																			: 0;
-																	const isDark =
-																		document.documentElement.classList.contains(
-																			"dark",
-																		);
-
-																	const baseAlpha = isDark ? 0.25 : 0.15;
-																	const scale = isDark ? 0.65 : 0.55;
-
-																	const alpha =
-																		value > 0
-																			? baseAlpha + intensity * scale
-																			: isDark
-																				? 0.08
-																				: 0.06;
-																	const isTopLeft =
-																		actualLabel === "no-damage" &&
-																		predictedLabel === "no-damage";
-																	const backgroundColor = isTopLeft
-																		? isDark
-																			? `rgba(5, 150, 105, ${Math.min(0.95, alpha + 0.25)})`
-																			: `rgba(16, 185, 129, ${alpha})`
-																		: isDiagonal
-																			? `rgba(16, 185, 129, ${alpha})`
-																			: `rgba(239, 68, 68, ${alpha})`;
-																	return (
-																		<td
-																			key={`cell-${actualLabel}-${predictedLabel}`}
-																			className={`px-2 py-2 text-center font-medium ${
-																				isDiagonal
-																					? "text-emerald-900 dark:text-emerald-300"
-																					: "text-rose-700 dark:text-rose-300"
-																			}`}
-																			style={{ backgroundColor }}
-																		>
-																			{value}
-																		</td>
-																	);
-																})}
-																<td className="px-2 py-2 text-center font-semibold text-muted-foreground">
-																	{rowTotal}
-																</td>
-															</tr>
-														);
-													})}
-												</tbody>
-											</table>
-										</div>
-									) : (
-										<div className="rounded-md border border-border bg-background px-4 py-6 text-sm text-muted-foreground">
-											No predicted VLM labels available yet to build the
-											confusion matrix.
-										</div>
-									)}
-								</div>
-							</div>
+						{dashboard.activeSection === "overview" && (
+							<DashboardOverviewSection
+								loadingStats={dashboard.loadingStats}
+								loadingBuildings={dashboard.loadingBuildings}
+								totalBuildings={dashboard.stats.total}
+								predictionMetrics={dashboard.predictionMetrics}
+								overviewDamageRows={dashboard.overviewDamageRows}
+								onViewBuildings={dashboard.setDamageFilter}
+							/>
 						)}
 
-						{activeSection === "buildings" && (
-							<div className="rounded-xl border border-border bg-card p-5 overflow-hidden relative">
-								<div className="mb-4 flex flex-wrap items-center justify-between gap-3">
-									<h3 className="text-lg font-semibold">Buildings</h3>
-									<div className="flex items-center gap-3">
-										<p className="text-xs text-muted-foreground">
-											{totalItems} total rows
-										</p>
-									</div>
-								</div>
-
-								<div className="mb-4 flex flex-wrap items-center justify-between gap-3">
-									<Select
-										value={
-											activeDamageFilter === "all"
-												? undefined
-												: activeDamageFilter
-										}
-										onValueChange={(value) => {
-											setActiveDamageFilter(value);
-											setPage(1);
-										}}
-									>
-										<SelectTrigger className="w-full max-w-48">
-											<SelectValue placeholder="Damages" />
-										</SelectTrigger>
-										<SelectContent>
-											<SelectGroup>
-												<SelectLabel>Damages</SelectLabel>
-												{DAMAGE_FILTERS.map((filter) => (
-													<SelectItem key={filter.key} value={filter.key}>
-														{filter.label}
-													</SelectItem>
-												))}
-											</SelectGroup>
-										</SelectContent>
-									</Select>
-
-									<div className="flex items-center gap-2">
-										<label
-											htmlFor="align-item"
-											className="text-sm text-muted-foreground"
-										>
-											Correct
-										</label>
-										<Switch
-											id="align-item"
-											checked={showCorrectOnly}
-											onCheckedChange={setShowCorrectOnly}
-										/>
-									</div>
-								</div>
-
-								{loadingBuildings ? (
-									<div className="rounded-md border border-border bg-background p-4">
-										{loadingBuildings && (
-											<div className="absolute inset-0 z-10 flex items-center justify-center bg-card/40 backdrop-blur-[2px]">
-												<SpinnerEmpty
-													title="Loading building records"
-													description={`Retrieving building rows for page ${page} with the ${prettyLabel(activeDamageFilter)} filter.`}
-													className="min-h-[280px] border-0 p-0"
-												/>
-											</div>
-										)}
-									</div>
-								) : (
-									<div className="relative z-0 overflow-x-auto">
-										<table className="w-full min-w-[980px] text-sm">
-											<thead>
-												<tr className="border-b border-border/60 text-left text-muted-foreground bg-muted/30">
-													<th className="py-2 pr-3 font-medium">Building</th>
-													<th className="py-2 pr-3 font-medium">Disaster</th>
-													<th className="py-2 pr-3 font-medium">xBD</th>
-													<th className="py-2 pr-3 font-medium">Actual</th>
-													<th className="py-2 pr-3 font-medium">Predicted</th>
-													<th className="py-2 pr-3 font-medium">Correct</th>
-												</tr>
-											</thead>
-											<tbody>
-												{rows.map((building) => {
-													const thumbnail = resolveStorageUrl(
-														building.post_image_path ?? building.pre_image_path,
-													);
-
-													return (
-														<tr
-															key={building.uid}
-															className="border-b border-border/70 align-top cursor-pointer hover:bg-muted/40"
-															onClick={() => {
-																if (!selectedDisasterName) return;
-																const buildingQuery = `building=${encodeURIComponent(building.uid)}`;
-																navigate(
-																	`/map/${selectedDisasterName}/${building.xbd_id}?${buildingQuery}`,
-																);
-															}}
-														>
-															<td className="py-3 pr-3">
-																<Item
-																	imageSrc={thumbnail}
-																	imageAlt={`Building ${building.uid}`}
-																	title={building.address || building.uid}
-																	subtitle={
-																		building.address ? building.uid : undefined
-																	}
-																/>
-															</td>
-															<td className="py-3 pr-3 font-medium text-xs whitespace-nowrap">
-																{building.disaster_name}
-															</td>
-															<td className="py-3 pr-3 text-muted-foreground">
-																{building.xbd_id}
-															</td>
-															<td className="py-3 pr-3">
-																<span
-																	className="inline-flex whitespace-nowrap rounded-md px-2 py-1 text-xs font-medium text-white"
-																	style={{
-																		backgroundColor:
-																			DAMAGE_COLOR_HEX[
-																				normalizeDamage(building.actual_damage)
-																			],
-																	}}
-																>
-																	{prettyLabel(
-																		normalizeDamage(building.actual_damage),
-																	)}
-																</span>
-															</td>
-															<td className="py-3 pr-3">
-																{building.predicted_damage == null ? (
-																	<span className="text-muted-foreground">
-																		-
-																	</span>
-																) : (
-																	<span
-																		className="inline-flex whitespace-nowrap rounded-md px-2 py-1 text-xs font-medium text-white"
-																		style={{
-																			backgroundColor:
-																				DAMAGE_COLOR_HEX[
-																					building.predicted_damage
-																				],
-																		}}
-																	>
-																		{prettyLabel(building.predicted_damage)}
-																	</span>
-																)}
-															</td>
-															<td className="py-3 pr-3 text-muted-foreground">
-																{typeof building.is_correct === "boolean"
-																	? building.is_correct
-																		? "Yes"
-																		: "No"
-																	: "-"}
-															</td>
-														</tr>
-													);
-												})}
-											</tbody>
-										</table>
-									</div>
-								)}
-
-								<div className="relative z-20 border-t border-border bg-card px-4 pb-4 pt-2">
-									<Pagination
-										page={page}
-										totalPages={totalPages}
-										onPageChange={setPage}
-									/>
-								</div>
-							</div>
+						{dashboard.activeSection === "buildings" && (
+							<DashboardBuildingsSection
+								totalItems={dashboard.totalItems}
+								activeDamageFilter={dashboard.activeDamageFilter}
+								showCorrectOnly={dashboard.showCorrectOnly}
+								onDamageFilterChange={dashboard.setBuildingsFilter}
+								onShowCorrectOnlyChange={dashboard.setShowCorrectOnly}
+								loadingBuildings={dashboard.loadingBuildings}
+								page={dashboard.page}
+								rows={dashboard.rows}
+								selectedDisasterName={dashboard.selectedDisasterName}
+								totalPages={dashboard.totalPages}
+								onPageChange={dashboard.setPage}
+							/>
 						)}
 
-						{activeSection === "image-pairs" && (
-							<div className="rounded-xl border border-border bg-card p-5 overflow-hidden relative">
-								<div className="mb-4 flex flex-wrap items-center justify-between gap-3">
-									<h3 className="text-lg font-semibold">Image Pairs</h3>
-									<div className="flex items-center gap-3">
-										<p className="text-xs text-muted-foreground">
-											{imagePairTotalItems} total rows
-										</p>
-									</div>
-								</div>
-
-								{loadingImagePairs ? (
-									<div className="rounded-md border border-border bg-background p-4">
-										<div className="absolute inset-0 z-10 flex items-center justify-center bg-card/40 backdrop-blur-[2px]">
-											<SpinnerEmpty
-												title="Loading image pairs"
-												description="Calculating image-level metrics..."
-												className="min-h-[280px] border-0 p-0"
-											/>
-										</div>
-									</div>
-								) : (
-									<div className="relative z-0 overflow-x-auto">
-										<table className="w-full text-sm">
-											<thead>
-												<tr className="border-b border-border/60 text-left text-muted-foreground bg-muted/30">
-													<th className="py-2 pr-3 font-medium">Post Image</th>
-													<th className="py-2 pr-3 font-medium">
-														Total Buildings
-													</th>
-													<th className="py-2 pr-3 font-medium">Correct</th>
-													<th className="py-2 pr-3 font-medium">Incorrect</th>
-													<th className="py-2 pr-3 font-medium">Accuracy</th>
-												</tr>
-											</thead>
-											<tbody>
-												{imagePairPageRows.map((pair) => {
-													const postThumbnail = resolveStorageUrl(
-														pair.post_image_path,
-													);
-
-													return (
-														<tr
-															key={pair.id}
-															className="border-b border-border/70 align-top cursor-pointer hover:bg-muted/40"
-															onClick={() => {
-																if (!selectedDisasterName) return;
-																navigate(
-																	`/map/${selectedDisasterName}/${pair.xbd_id}`,
-																);
-															}}
-														>
-															<td className="py-3 pr-3">
-																<Item
-																	imageSrc={postThumbnail}
-																	imageAlt={`Post scene ${pair.xbd_id}`}
-																	title={`Scene #${pair.xbd_id}`}
-																	subtitle={`Pair ${pair.id}`}
-																	meta={`Compared: ${pair.comparedCount}`}
-																/>
-															</td>
-															<td className="py-3 pr-3 font-medium">
-																{pair.totalBuildings}
-															</td>
-															<td className="py-3 pr-3 text-emerald-600 font-medium">
-																{pair.correctCount}
-															</td>
-															<td className="py-3 pr-3 text-rose-600 font-medium">
-																{pair.incorrectCount}
-															</td>
-															<td className="py-3 pr-3 text-muted-foreground">
-																{pair.accuracyPct == null
-																	? "N/A"
-																	: `${pair.accuracyPct}%`}
-															</td>
-														</tr>
-													);
-												})}
-											</tbody>
-										</table>
-										{imagePairRows.length === 0 && (
-											<div className="rounded-md border border-border bg-background px-4 py-6 text-sm text-muted-foreground">
-												No image pair data available for this disaster yet.
-											</div>
-										)}
-									</div>
-								)}
-
-								<div className="relative z-20 border-t border-border bg-card px-4 pb-4 pt-2">
-									<Pagination
-										page={imagePairPageClamped}
-										totalPages={imagePairTotalPages}
-										onPageChange={setImagePairsPage}
-									/>
-								</div>
-							</div>
+						{dashboard.activeSection === "image-pairs" && (
+							<DashboardImagePairsSection
+								loadingImagePairs={dashboard.loadingImagePairs}
+								imagePairTotalItems={dashboard.imagePairTotalItems}
+								imagePairRows={dashboard.imagePairRows}
+								imagePairPageRows={dashboard.imagePairPageRows}
+								imagePairPage={dashboard.imagePairPage}
+								imagePairTotalPages={dashboard.imagePairTotalPages}
+								selectedDisasterName={dashboard.selectedDisasterName}
+								onPageChange={dashboard.setImagePairsPage}
+							/>
 						)}
 					</div>
 				</div>
